+++ identity-wallet/setler-cli/src/actions/index.js
@@ -4,7 +4,6 @@ import * as auth from "./auth.js";
 import * as config from "./config.js";
 import * as echo from "./echo.js";
 import * as escrow from "./escrow.js";
-import * as ident from "./ident.js";
 import * as kudos from "./kudos.js";
 import * as ledger from "./ledger.js";
 import * as message from "./message.js";
@@ -19,7 +18,6 @@ export {
   config,
   echo,
   escrow,
-  ident,
   kudos,
   ledger,
   message,