+++ identity-wallet/setler-cli/src/kudos.js
@@ -1,7 +1,4 @@
 // creatorPool is an array of objects with the following properties:
 //  id: string
 //  weight: number 0-1 (optional), will be scaled according to application, but 1 = full, 0 = none
-export const creatorPool = [
-  { id: "did:kudos:email:setler@loremlabs.com", weight: 1 },
-  //    {id: 'did:kudos:email:mankins+3@gmail.com', weight: .1},
-];
+export const creatorPool = [{ id: "email:setler@loremlabs.com", weight: 1 }];